19: Moving WhoAmI component to folder structure
